Okay, this is really jank, but it'll work.
So for users right now, you athenticate by sending an id along with your request as the body parameter 'token' to the /api/v1 endpoints.

The only endpoints not protected by this are /api/v1/auth/facebook*.

Call localhost:8081/api/v1/auth/facebook to login with facebook. This will redirect the user back to localhost:8081/ with the url parameter token=...

That ... is the token you can send to the api each time to authenticate. Save it in a cookie or something on the frontend and don't let the user see it I guess.