Added a bunch of google credential code to refresh access tokens and crap. No idea if it works, but there's like a 50% chance it will.
Also implemented a get_user_calendar function. Given a user id for a user who has authorized their calendar, it should return the next ten events on their calendar in an array. It will return an empty array in the callback if it failed. Or null if the user hasn't authorized google.